refactor: extract helpers from the time picker connector
